[CO] as_dict with list of pyleecan objects
