+++ pyleecan/Generator/ClassesRef/Simulation/DataKeeper.csv
@@ -1,5 +1,5 @@
 Variable name,Unit,Description (EN),Size,Type,Default value,Minimum value,Maximum value,,Package,Inherit,Methods,Constant Name,Constant Value,Class description
-name,-,Data name,0,str,,,,,Simulation,,,VERSION,1,Class for defining data to keep on a multi-simulation
+name,-,Data name,0,str,,,,,Simulation,,as_dict,VERSION,1,Class for defining data to keep on a multi-simulation
 symbol,-,Data symbol,,str,,,,,,,,,,
 unit,-,Data unit,,str,,,,,,,,,,
 keeper,-,Function that takes an Output in argument and return a value,,function,None,,,,,,,,,
